New module for associating files with phrases - assocfile (#136)
* New module for associating files with phrases - assocfile

Based on "associate" module, lets associate files like images or videos
to be sent into a chat with phrases. It's looking for files inside
"files" folder in Kassy's main directory.

* One small fix, forgot to remove this line

* Little bug fixes, now it 100% works

It didn't want to send files with sendFile to Facebook, sendImage works
for video and audio files as well, but it didn't work for text file I
tried to send for some reason... Additionally corrected some messages
output.